+++ changelog.md
@@ -93,6 +93,7 @@ Codion Change Log
 - ButtonPanelBuilder, overloaded builder() methods removed.
 - MenuBuilder, overloaded builder() methods removed.
 - MenuItemBuilder, overloaded builder() methods removed.
+- ToolBarBuilder, overloaded builder() methods removed.
 ### is.codion.swing.framework.model
 - EntityComboBoxModel.Builder.EntityTypeBuilder and ConnectionProviderBuilder added.
 ### is.codion.swing.framework.ui